Single-source the host-side CUDA RAII owners into src/cuda_raii.h
frame.cpp and stream.cpp each carried their own copy of the same host-side
CUDA glue: grow-only device/pinned buffers and stream/event owners, plus a
cudaSuccess-check macro. Consolidate the owners and the check into one internal
header both translation units include, so the next CUDA host TU reuses them
instead of re-copying.

The issue was filed against an earlier stream.cpp; the two copies had since
diverged, so this is not a verbatim move - I unify to the strictly-safer form
and keep each site's failure policy:

- The owners moved are the grow-only DevBuf/PinnedBuf (with ensure(), a tracked
  cap, and an explicit null-on-failure) plus StreamOwner/EventOwner, under the
  new named namespace cudec_cuda. frame.cpp dropped its one-shot DevPtr and now
  uses cudec_cuda::DevBuf: every frame allocation is a single ensure() from
  cap 0 over a non-zero size (guarded by n != 0), which allocates exactly once
  - identical to the old cudaMalloc, and DevBuf additionally nulls the pointer
  on a failed allocation, so it is strictly safer on the reject path.

- The check macro is single-sourced as CUDEC_CUDA_CHECK(call, on_failure): it
  hoists the `!= cudaSuccess` test while leaving the failure action to each
  site. frame's FRAME_CUDA stays a plain `return CUDEC_ERR_CUDA`; stream's GROW
  keeps poisoning the reusable context before returning. Both are now thin
  wrappers over the shared macro; the two were no longer byte-identical (GROW
  poisons), so a plain shared macro would have changed behaviour.

The inline `!= cudaSuccess` / WAVE_FAIL break-path checks in the stream wave
loop (which must break, not return) and MulOverflows stay in stream.cpp.

Lock the property with a configure-time drift detector in tests/CMakeLists.txt:
the build reds if cuda_raii.h stops defining an owner, or if a CUDA host TU
drops the include or re-introduces a local owner struct. Same drift-detector
scope limit as the existing structural checks - it catches a re-copy under the
same names, not one under a fresh name.

No ABI, kernel, or parser change; behaviour is bit-identical. Verified in the
pinned CUDA container: 10/10 ctest passed, -Werror clean, frame_twin and
stream_twin (the behaviour-preservation pins) green.

Closes #40
